requirements/docs.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.8
django-coreapi==2.3.0
django-cors-headers==3.1.0
django-mysql==3.2.0
djangorestframework==3.10.2
et-xmlfile==1.0.1
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
MarkupSafe==1.1.1
memory-profiler==0.55.0
mysqlclient==1.4.4
openpyxl==2.6.0
psutil==5.6.3
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
